requirements.txt
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
python-ldap==3.4.4
